Split naming module off of utils
Triggered by circular import exceptions that occur when utils is
imported before typing (but not the other way around). Having a utils
module is a code smell anyway.
